Saving and Loading Files
Saving and Loading to/from files has been implemented. Files are identified with the ".q" extension. As of right now, these are put in the SavedItems folder, but this can be changed to whatever is easiest.
Some QOL improvements to add: Overwrite Menu (popup), Deletion Confirmation (Popup), Different Icon Layout (current one is kind of ugly).